Refactor a lot of messaging.
Refactor messaging, parallel with the server application, to be able to
remove a lot of components that originated from the server application.
Event validation will be done fully on the server-side, making the client
dumb. This allows us to easier maintain the software systems in parallel.

This commit is intentionally completely broken, sorry about that. I have
to save my progress so far in order to be able to merge Floris his latest
changes into my branch.