Implemented transition to Lobby activity and has a game state as a boolean.
Whenever the game state is 'false' the lobby should be started. Whenever this activity is started again from another activity it will retrieve a boolean in the onCreate method if the extras from the intent is not null. This way the activity knows if the game has started already and if it should start the lobby for a new game or not.